refactor: reuse sequence entries by including superset in policy_definition model
