+++ vmngclient/endpoints/configuration_policy_data_definition_builder.py
@@ -1,162 +1,41 @@
 # mypy: disable-error-code="empty-body"
-from enum import Enum
-from typing import Any, List, Optional, Union
+from typing import List, Optional, Union
 
-from pydantic import BaseModel, Field, IPvAnyNetwork
-from typing_extensions import Annotated, Literal
+from pydantic import BaseModel, Field
+from typing_extensions import Annotated
 
 from vmngclient.endpoints import APIEndpoints, delete, get, post, put
 from vmngclient.model.policy.policy_definition import (
+    AppListEntry,
+    DefaultAction,
+    DefaultActionType,
+    DestinationDataIPv6PrefixListEntry,
+    DestinationDataPrefixListEntry,
+    DestinationIPEntry,
+    DestinationPortEntry,
+    DestinationRegionEntry,
+    DNSAppListEntry,
+    DNSEntry,
+    DSCPEntry,
+    PacketLengthEntry,
+    PLPEntry,
     PolicyDefinition,
     PolicyDefinitionCreationPayload,
     PolicyDefinitionEditPayload,
     PolicyDefinitionEditResponse,
     PolicyDefinitionId,
     PolicyDefinitionPreview,
+    ProtocolEntry,
+    SourceDataIPv6PrefixListEntry,
+    SourceDataPrefixListEntry,
+    SourceIPEntry,
+    SourcePortEntry,
+    TCPEntry,
+    TrafficToEntry,
 )
 from vmngclient.typed_list import DataSequence
 
-# TODO: add validators for custom strings (eg.: port ranges, space separated networks)
-# TODO: model actions
-
-
-class DefaultActionType(str, Enum):
-    DROP = "drop"
-    ACCEPT = "accept"
-
-
-class PLPEntryValues(str, Enum):
-    LOW = "low"
-    HIGH = "high"
-
-
-class DNSEntryValues(str, Enum):
-    REQUEST = "request"
-    RESPONSE = "response"
-
-
-class TrafficToEntryValues(str, Enum):
-    ACCESS = "access"
-    CORE = "core"
-    SERVICE = "service"
-
-
-class DestinationRegionEntryValues(str, Enum):
-    PRIMARY = "primary-region"
-    SECONDARY = "secondary-region"
-    OTHER = "other-region"
-
-
-class SequenceIpType(str, Enum):
-    IPV4 = "ipv4"
-    IPV6 = "ipv6"
-    ALL = "all"
-
-
-class BaseAction(str, Enum):
-    DROP = "drop"
-    ACCEPT = "accept"
-
-
-class SequenceType(str, Enum):
-    APPLICATION_FIREWALL = "applicationFirewall"
-    DATA = "data"
-    SERVICE_CHAINING = "serviceChaining"
-    TRAFFIC_ENGINEERING = "trafficEngineering"
-    QOS = "qos"
-
-
-class PacketLengthEntry(BaseModel):
-    field: Literal["packetLength"]
-    value: str = Field(description="0-65536 range or single number")
-
-
-class PLPEntry(BaseModel):
-    field: Literal["plp"]
-    value: PLPEntryValues
-
-
-class ProtocolEntry(BaseModel):
-    field: Literal["protocol"]
-    value: str = Field(description="0-255 single numbers separate by space")
-
-
-class DSCPEntry(BaseModel):
-    field: Literal["dscp"]
-    value: str = Field(description="0-63 single numbers separate by space")
-
-
-class SourceIPEntry(BaseModel):
-    field: Literal["sourceIp"]
-    value: str = Field(description="IP network specifier separate by space")
-
-
-class SourcePortEntry(BaseModel):
-    field: Literal["sourcePort"]
-    value: str = Field(description="0-65535 range or separate by space")
-
-
-class DestinationIPEntry(BaseModel):
-    field: Literal["destinationIp"]
-    value: IPvAnyNetwork
-
-
-class DestinationPortEntry(BaseModel):
-    field: Literal["destinationPort"]
-    value: str = Field(description="0-65535 range or separate by space")
-
-
-class TCPEntry(BaseModel):
-    field: Literal["tcp"]
-    value: Literal["syn"]
-
-
-class DNSEntry(BaseModel):
-    field: Literal["dns"]
-    value: DNSEntryValues
-
-
-class TrafficToEntry(BaseModel):
-    field: Literal["trafficTo"]
-    value: TrafficToEntryValues
-
-
-class DestinationRegionEntry(BaseModel):
-    field: Literal["destinationRegion"]
-    value: DestinationRegionEntryValues
-
-
-class SourceDataPrefixListEntry(BaseModel):
-    field: Literal["sourceDataPrefixList"]
-    ref: str
-
-
-class DestinationDataPrefixListEntry(BaseModel):
-    field: Literal["destinationDataPrefixList"]
-    ref: str
-
-
-class SourceDataIPv6PrefixListEntry(BaseModel):
-    field: Literal["sourceDataIpv6PrefixList"]
-    ref: str
-
-
-class DestinationDataIPv6PrefixListEntry(BaseModel):
-    field: Literal["destinationDataIpv6PrefixList"]
-    ref: str
-
-
-class DNSAppListEntry(BaseModel):
-    field: Literal["dnsAppList"]
-    ref: str
-
-
-class AppListEntry(BaseModel):
-    field: Literal["appList"]
-    ref: str
-
-
-Entry = Annotated[
+DataPolicySequenceEntry = Annotated[
     Union[
         PacketLengthEntry,
         PLPEntry,
@@ -181,35 +60,13 @@ Entry = Annotated[
 ]
 
 
-class Match(BaseModel):
-    entries: List[Entry]
-
-
-class Action(BaseModel):
-    pass
-
-
-class Sequence(BaseModel):
-    sequence_id: int = Field(alias="sequenceId")
-    sequence_name: str = Field(alias="sequenceName")
-    base_action: BaseAction = Field(alias="baseAction")
-    sequence_type: SequenceType = Field(alias="sequenceType")
-    sequence_ip_type: SequenceIpType = Field(alias="sequenceIpType")
-    match: Match
-    actions: List[Any]
-
-
-class DefaultAction(BaseModel):
-    type: DefaultActionType
-
-
 class Data(BaseModel):
     type: str = Field(default="data", const=True)
     default_action: Optional[DefaultAction] = Field(
         default=DefaultAction(type=DefaultActionType.DROP), alias="defaultAction"
     )
     is_activated_by_vsmart: Optional[bool] = Field(default=False, alias="isActivatedByVsmart")
-    sequences: List[Sequence] = []
+    sequences: List[DataPolicySequenceEntry] = []
 
 
 class DataDefinitionCreationPayload(Data, PolicyDefinitionCreationPayload):
